+++ intellij_project/src/main/java/it/polimi/telco_webapp/entities/OptionsPricesView.java
@@ -6,6 +6,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import org.eclipse.persistence.annotations.ReadOnly;
 
+import java.math.BigDecimal;
+
 
 @Entity
 @Table(name = "options_ordered_prices", schema = "telco_db")
@@ -15,9 +17,9 @@ public class OptionsPricesView {
     @Id
     private int order_id;
     @Column
-    private float sum_of_options;
+    private BigDecimal sum_of_options;
 
     public int getOrderId(){return order_id;}
-    public float getSumOfOptionsPurchased(){return sum_of_options;}
+    public BigDecimal getSumOfOptionsPurchased(){return sum_of_options;}
     /* Since this is a view, the data is read-only. No setters. */
 }
